refactor(reader): extract Strong's document builder
Why:
- BibleReaderController still owned Strong's dictionary lookup, morphology resolution, linkification, and multi-fragment document assembly.
- Issue #146 requires moving non-orchestration responsibilities into focused collaborators without changing Android parity behavior.

What Changed:
- Added BibleReaderStrongsDocumentBuilder to own Strong's and morphology MultiDocument payload construction, dictionary key validation, linkification, curated dictionary selection, and missing-dictionary fallback generation.
- Reduced BibleReaderController to dependency wiring and bridge routing for Strong's results while preserving restored Multi document and word lookup behavior through the shared lookup helper.
- Moved Strong's helper tests to assert the new builder ownership boundary.

Validation:
- DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild test -project AndBible.xcodeproj -scheme AndBible -destination 'platform=iOS Simulator,name=iPhone 17,OS=26.5' -only-testing:AndBibleTests/AndBibleTests/testStrongsLookupKeyOptionsIncludeIntermediateZeroTrimVariants
- DEVELOPER_DIR=/Applications/Xcode.app/Contents/Developer xcodebuild test -project AndBible.xcodeproj -scheme AndBible -destination 'platform=iOS Simulator,name=iPhone 17,OS=26.5' with focused Strong's/dictionary/link-window tests
- python3 scripts/check_repo_standards.py docblocks --all-files
- npx gitnexus detect-changes --repo /Users/primetheus/Desktop/and-bible-convert/and-bible-ios/.worktrees/issue-146-strongs-document-builder

Impact:
- Preserves Android-style Strong's MultiDocument routing and missing-dictionary fallback while shrinking the controller by moving dictionary document ownership into a focused collaborator.
- Word lookup remains on its existing controller path but shares the same dictionary validation helper.

Refs:
- #146
